12_21

+++ src/StateTask.cpp
@@ -748,7 +748,7 @@ bool StateTask::CheckMaxonPosition(std::shared_ptr<MaxonMotor> motor)
             return false;
         }
 
-        std::tuple<int, float, int> parsedData = MParser.parseRecieveCommand(&frame);
+        std::tuple<int, float, float> parsedData = MParser.parseRecieveCommand(&frame);
         motor->currentPos = std::get<1>(parsedData);
         std::cout << "Current Position of [" << std::hex << motor->nodeId << std::dec << "] : " << motor->currentPos << endl;
         return true;
